places api with retrofit done

+++ app/src/main/java/com/codingblocks/groupchat/places/Result.java
@@ -3,10 +3,6 @@ package com.codingblocks.groupchat.places;
 
 import java.util.List;
 
-import com.codingblocks.groupchat.Geometry;
-import com.codingblocks.groupchat.OpeningHours;
-import com.codingblocks.groupchat.Photo;
-import com.codingblocks.groupchat.places.*;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
@@ -35,7 +31,7 @@ public class Result {
     private String placeId;
     @SerializedName("rating")
     @Expose
-    private int rating;
+    private double rating;
     @SerializedName("reference")
     @Expose
     private String reference;
@@ -108,7 +104,7 @@ public class Result {
         this.placeId = placeId;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
